requirements.txt
# Discord Bot Dependencies for Railway
discord.py>=2.3.0
aiohttp>=3.9.0
supabase>=2.0.0
python-dotenv>=1.0.0
